+++ app/Http/Controllers/LogotypeController.php
@@ -19,7 +19,9 @@ class LogotypeController extends Controller
     public function LogotypeDeleteSubmit($id)
     {
         $logotypeItem = Logotype::find($id);
-        unlink(public_path(Storage::url($logotypeItem->path_to_file)));
+        if (file_exists(public_path(Storage::url($logotypeItem->path_to_file)))) {
+            unlink(public_path(Storage::url($logotypeItem->path_to_file)));
+        }
         Logotype::find($id)->delete();
         return redirect()->route('logotype')->with('success', 'Логотип удален');
     }
